refactor diff and pro functions to use smaller integer types

+++ scalar/src/lib.rs
@@ -2,11 +2,11 @@ pub fn sum(a: u8, b: u8) -> u8 {
     return a + b;
 }
 
-pub fn diff(a: i32, b: i32) -> i32 {
+pub fn diff(a: i16, b: i16) -> i16 {
     return a - b;
 }
 
-pub fn pro(a: i32, b: i32) -> i32 {
+pub fn pro(a: i8, b: i8) -> i8 {
     return a * b;
 }
